+++ tests/__init__.py
@@ -1,3 +1,25 @@
-from tests.base import Page, Component, element_attribute_not_to_include
+from tests.common import (
+    Component,
+    Page,
+    has_element,
+    element_attribute_not_to_include,
+    Player,
+    Albums,
+    Topbar,
+    TopArtists,
+    Tracks,
+    Sidebar,
+)
 
-__all__ = (Page, Component, element_attribute_not_to_include)
+__all__ = (
+    Component,
+    Page,
+    has_element,
+    element_attribute_not_to_include,
+    Player,
+    Albums,
+    Topbar,
+    TopArtists,
+    Tracks,
+    Sidebar,
+)
