+++ tests/favorites_test.py
@@ -1,13 +1,12 @@
 import os
 import unittest
 
-import selenium
 from selenium.webdriver import DesiredCapabilities, Remote
 from selenium.webdriver.support.ui import WebDriverWait
 from selenium.webdriver.common.by import By
 from selenium.webdriver.support import expected_conditions as EC
 
-from tests import Page, Component, element_attribute_not_to_include
+from tests import Page, Tracks
 from tests.login_test import LoginPage
 from tests.main_test import MainPage
 
@@ -17,44 +16,7 @@ class FavoritesPage(Page):
 
     @property
     def track_list(self):
-        return TrackList(self.driver)
-
-
-class TrackList(Component):
-    ALBUM_IMG = '//img[@class="track__artwork__img"]'
-    ARTIST_LBL = '//div[@class="track__container__artist"]'
-    TRACK_FAV_BTN = '//img[@class="track-fav"]'
-    TRACK_FAV_BTN_BY_ID = '//img[@class="track-fav" and @data-id="{}"]'
-
-    def get_first_track_id(self):
-        fav_btn = self.driver.find_element_by_xpath(self.TRACK_FAV_BTN)
-        return fav_btn.get_attribute("data-id")
-
-    def open_album(self):
-        self.driver.find_element_by_xpath(self.ALBUM_IMG).click()
-
-    def open_artist(self):
-        self.driver.find_element_by_xpath(self.ARTIST_LBL).click()
-
-    def get_fav_btn(self, track_id):
-        return self.driver.find_element_by_xpath(self.TRACK_FAV_BTN_BY_ID.format(track_id))
-
-    def remove_favor(self, track_id):
-        self.get_fav_btn(track_id).click()
-        WebDriverWait(self.driver, 10).until(
-            element_attribute_not_to_include((By.XPATH, self.TRACK_FAV_BTN_BY_ID.format(track_id)), "data-in_favorites")
-        )
-
-    def add_to_favorites(self, track_id):
-        self.get_fav_btn(track_id).click()
-        WebDriverWait(self.driver, 10).until(
-            EC.element_attribute_to_include((By.XPATH, self.TRACK_FAV_BTN_BY_ID.format(track_id)), "data-in_favorites")
-        )
-
-    def track_is_in_favor(self, track_id):
-        return bool(
-            self.get_fav_btn(track_id).get_attribute("data-in_favorites")
-        )
+        return Tracks(self.driver)
 
 
 class FavoritesTest(unittest.TestCase):
@@ -89,13 +51,13 @@ class FavoritesTest(unittest.TestCase):
         self.driver.quit()
 
     def test_album_opening(self):
-        self.favorites_page.track_list.open_album()
+        self.favorites_page.track_list.open_first_album()
         WebDriverWait(self.driver, 10).until(
             EC.presence_of_element_located((By.CLASS_NAME, "album"))
         )
 
     def test_artist_opening(self):
-        self.favorites_page.track_list.open_artist()
+        self.favorites_page.track_list.open_first_artist()
         WebDriverWait(self.driver, 10).until(
             EC.presence_of_element_located((By.CLASS_NAME, "artist"))
         )
@@ -116,4 +78,8 @@ class FavoritesTest(unittest.TestCase):
 
     def test_favorites_opening_from_navbar(self):
         self.main_page = MainPage(self.driver)
-        self.main_page.navbar.open_favorites()
+        self.main_page.sidebar.go_favorites()
+        WebDriverWait(self.driver, 10).until(
+            EC.presence_of_element_located((By.CLASS_NAME, "favorites__description-title"))
+        )
+
